Apply the polish review fixes: one look, readable errors
No stock Material control is left in the app: every text box, chip,
checkbox, switch, text button and divider is now the square house
version, including inside the capability dialog, the view editor and
the merge screen. Stop sits level over Cancel, and dialogs spring in.

Errors read in bold ink beside a madder bar, with the field outlined in
madder, and tests pin both contrasts. Chips tell TalkBack what they do:
a single choice is a radio button, a toggle is a checkbox, and an action
is a button. A field's hint or error is read once, with the field.

Song rows show the title over the artist, as the session list does.
Repeated button words and sentences now live in Messages, and the test
harnesses share one way into each route.

572 shared + 78 app JVM + 213 instrumented tests, 0 failures.

+++ shared/src/commonMain/kotlin/dev/repertosaurus/session/TimedHistory.kt
@@ -17,7 +17,7 @@ public data class TimedHistory(val events: List<TimedEvent>) {
     /** Journal session 11 D85 #1: the song detail lists the newest [SHOWN] and counts the rest. */
     val latest: List<TimedEvent> get() = events.take(SHOWN)
 
-    val olderCount: Int get() = (events.size - SHOWN).coerceAtLeast(0)
+    val olderCount: Long get() = (events.size - SHOWN).coerceAtLeast(0).toLong()
 
     public companion object {
         public const val SHOWN: Int = 10
